Move response class to integration.http package MODPATRON-66
